extract RichTextEditor into own directory
