+++ src/components/formBuilder/widgets/RichTextWidget.tsx
@@ -17,53 +17,10 @@
 
 import React from "react";
 import { type WidgetProps } from "@rjsf/utils";
-import { useCurrentEditor, EditorProvider } from "@tiptap/react";
-import { StarterKit } from "@tiptap/starter-kit";
-import { Button, ButtonGroup, ButtonToolbar } from "react-bootstrap";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBold, faItalic } from "@fortawesome/free-solid-svg-icons";
-import styles from "./RichTextWidget.module.scss";
-
-const Toolbar: React.FunctionComponent = () => {
-  const { editor } = useCurrentEditor();
-
-  if (!editor) {
-    return null;
-  }
-
-  return (
-    <ButtonToolbar className={styles.toolbar}>
-      <ButtonGroup size="sm" aria-label="Rich-Text Editor Toolbar">
-        <Button
-          variant="default"
-          onClick={() => editor.chain().focus().toggleBold().run()}
-          disabled={!editor.can().chain().focus().toggleBold().run()}
-          active={editor.isActive("bold")}
-        >
-          {/* TODO: Fix having to explicitly set height and width for document renderer */}
-          <FontAwesomeIcon icon={faBold} height="16" width="16" />
-        </Button>
-        <Button
-          variant="default"
-          onClick={() => editor.chain().focus().toggleItalic().run()}
-          disabled={!editor.can().chain().focus().toggleItalic().run()}
-          active={editor.isActive("italic")}
-        >
-          <FontAwesomeIcon icon={faItalic} height="16" width="16" />
-        </Button>
-      </ButtonGroup>
-    </ButtonToolbar>
-  );
-};
+import RichTextEditor from "@/components/richTextEditor/RichTextEditor";
 
 const RichTextWidget: React.FunctionComponent<WidgetProps> = () => (
-  <div className={styles.root}>
-    <EditorProvider
-      extensions={[StarterKit]}
-      content="<p>Hello TipTap! 🍌</p>"
-      slotBefore={<Toolbar />}
-    />
-  </div>
+  <RichTextEditor />
 );
 
 export default RichTextWidget;
